generates new guid on form submit
At the moment the form stays on the same page and doesn't redirect
to another page. If the user wants to submit another event they
would have the same guid as the last event. So, I have made it that
when a form is submitted, a new guid is generated.

If the form is submitted twice, each one should have a unique guid